requirements.txt
Flask==2.3.3
fer==22.5.1
opencv-python==4.8.1.78
tensorflow==2.13.0
Pillow==10.2.0
numpy==1.24.3
moviepy==1.0.3